requirements.txt
# Core
streamlit

# Streamlit components
streamlit-ace
streamlit-discourse
streamlit-disqus
streamlit-elements
streamlit-pandas-profiling
streamlit-player
streamlit-quill

# Dependencies
pandas-profiling
requests

Pillow==9.5.0
